build function empty list, add page empty list, and build terniary condition
