+++ src/pages/ListEmpty/ListEmpty.css
@@ -1,0 +1,3 @@
+button {
+  cursor: pointer;
+}
